Renamed file from articles_contianer to ArticlesContainer
